myproject/requirements.txt
django-silk
djangorestframework
djangorestframework-simplejwt